+++ src/libs/apis/voiceFile.ts
@@ -1,6 +1,9 @@
 import client from '@apis/client';
-import {ResultResponseData} from '@type/common';
-import {VoiceFileResponseData, VoiceFilesRequestData} from '@type/voiceFile';
+import {ResultResponseData} from '@type/api/common';
+import {
+  VoiceFileResponseData,
+  VoiceFilesRequestData,
+} from '@type/api/voiceFile';
 
 const getVoiceFiles = async ({alarmId}: Readonly<VoiceFilesRequestData>) => {
   const res = await client.get<ResultResponseData<VoiceFileResponseData>>(
