+++ src/libs/hooks/alarm/useGetAlarmCategoryByAlarmCategoryId.ts
@@ -1,6 +1,6 @@
 import {getAlarmCategoryByAlarmCategoryId} from '@apis/alarm';
 import {useQuery} from '@tanstack/react-query';
-import {AlarmCategoryRequestData} from '@type/alarm';
+import {AlarmCategoryRequestData} from '@type/api/alarm';
 
 const useGetAlarmCategoryByAlarmCategoryId = ({
   alarmCategoryId,
