+++ src/libs/hooks/voiceFile/useGetVoiceFiles.ts
@@ -1,6 +1,6 @@
 import {getVoiceFiles} from '@apis/voiceFile';
 import {useQuery} from '@tanstack/react-query';
-import {VoiceFilesRequestData} from '@type/voiceFile';
+import {VoiceFilesRequestData} from '@type/api/voiceFile';
 
 const useGetVoiceFiles = ({alarmId}: Readonly<VoiceFilesRequestData>) => {
   return useQuery({
