+++ src/screens/Letter/LetterList/index.tsx
@@ -10,8 +10,8 @@ import Txt from '@components/atom/Txt';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import {NativeStackScreenProps} from '@react-navigation/native-stack';
 import LetterCard from '@screens/Letter/LetterList/components/LetterCard';
-import {LetterStackParamList} from '@type/LetterStackParamList';
-import {LetterResponseData} from '@type/providedFile';
+import {LetterStackParamList} from '@type/nav/LetterStackParamList';
+import {LetterResponseData} from '@type/api/providedFile';
 import {useEffect, useState} from 'react';
 import {Alert, Pressable, Text, View} from 'react-native';
 import {ScrollView} from 'react-native-gesture-handler';
