+++ src/types/api/alarm.ts
