+++ src/types/component/TxtType.ts
